move `wraps` from `api_util` to `util` to avoid cyclic dependencies
