+++ jax/interpreters/sharded_jit.py
@@ -26,9 +26,9 @@ from . import xla
 from .. import linear_util as lu
 from ..lib import xla_bridge as xb
 from ..lib import xla_client as xc
-from ..api_util import flatten_axes, flatten_fun, wraps
+from ..api_util import flatten_axes, flatten_fun
 from ..tree_util import tree_flatten, tree_unflatten
-from ..util import extend_name_stack, wrap_name, safe_zip
+from ..util import extend_name_stack, wrap_name, wraps, safe_zip
 from ..config import config
 
 xops = xc._xla.ops
